Visualizing the network of consecutive replies in 'WhatsApp' chat logs

Plots a network for replies between authors in chat logs. Each message is evaluated as a reply to the previous one.

plot_network(
  data,
  names = "all",
  starttime = "1960-01-01 00:00",
  endtime = "2200-01-01 00:00",
  return_data = FALSE,
  collapse_sessions = FALSE,
  edgetype = "n",
  exclude_sm = FALSE
)

Arguments

data

A 'WhatsApp' chatlog that was parsed with parse_chat.

names

A vector of author names that the visualization will be restricted to. Non-listed authors will be removed.

starttime

Datetime that is used as the minimum boundary for exclusion. Is parsed with as.POSIXct. Standard format is "yyyy-mm-dd hh:mm". Is interpreted as UTC to be compatible with 'WhatsApp' timestamps.

endtime

Datetime that is used as the maximum boundary for exclusion. Is parsed with as.POSIXct. Standard format is "yyyy-mm-dd hh:mm". Is interpreted as UTC to be compatible with 'WhatsApp' timestamps.

return_data

If TRUE, returns a data frame of subsequent interactions with senders and recipients. Default is FALSE.

collapse_sessions

Whether multiple subsequent messages by the same sender should be collapsed into one row. Default is FALSE.

edgetype

What type of content is displayed as an edge. Must be one of "TokCount","EmojiCount","SmilieCount","LocationCount","URLCount","MediaCount" or "n".

exclude_sm

If TRUE, excludes the 'WhatsApp' system messages from the descriptive statistics. Default is FALSE.

Value

A network visualization of authors in 'WhatsApp' chat logs where each subsequent message is considered a reply to the previous one. Input will be ordered by TimeOrder column.

Examples

data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_network(data)