Restricting chat logs to certain authors or timeframes.

Excluding parts of the chat by senders or timestamps

tailor_chat(
  data,
  names = "all",
  starttime = "1960-01-01 00:00",
  endtime = "2200-01-01 00:00",
  exclude_sm = FALSE
)

Arguments

data

A 'WhatsApp' chat log that was parsed with parse_chat.

names

A vector of names that the output is restricted to. Messages from other non-contained authors are excluded.

starttime

Datetime that is used as the minimum boundary for exclusion. Is parsed with as.POSIXct. Standard format is "yyyy-mm-dd hh:mm". Is interpreted as UTC to be compatible with 'WhatsApp' timestamps.

endtime

Datetime that is used as the maximum boundary for exclusion. Is parsed with as.POSIXct. Standard format is "yyyy-mm-dd hh:mm". Is interpreted as UTC to be compatible with 'WhatsApp' timestamps.

exclude_sm

If TRUE, excludes the 'WhatsApp' system messages from the descriptive statistics. Default is FALSE.

Value

A dataframe that is restricted to the specified timeframe and authors

Examples

data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
tailor_chat(data, names = c("Mallory", "Alice"))
#>               DateTime Anonymous  Sender
#> 4  2018-01-29 12:24:00  Person_1 Mallory
#> 11 2018-01-29 23:33:00  Person_3   Alice
#> 15 2018-01-29 23:46:00  Person_3   Alice
#> 19 2018-01-29 23:47:00  Person_3   Alice
#> 21 2018-01-29 23:47:00  Person_3   Alice
#> 23 2018-01-29 23:49:00  Person_3   Alice
#> 25 2018-01-29 23:51:00  Person_3   Alice
#> 28 2018-01-29 23:53:00  Person_3   Alice
#> 29 2018-01-29 23:58:00  Person_3   Alice
#> 30 2018-01-29 23:59:00  Person_3   Alice
#> 33 2018-01-29 23:59:00  Person_3   Alice
#> 37 2018-01-30 00:00:00  Person_3   Alice
#> 40 2018-01-30 00:10:00  Person_3   Alice
#> 48 2018-01-30 00:14:00  Person_3   Alice
#>                                                                                                                                                                                                                                                                                                                                                                                                                                  Message
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                Hey! :)
#> 11                                                                                                                                                                                                                                                                                                                                                                                                         Ist das hier ein Chatverlauf?
#> 15                                                                                                                                                                                                                                                                                                                                                                                            PTT-20180129-WA0025.opus (Datei angehängt)
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                                  ja😄
#> 21                                                                                                                                                                                                                                                                                                                                                                                                                und Emojis gehen auch!
#> 23                                                                                                                                                                                                                                                                                                                                                                                      Keine Ahnung, lass uns das doch mal ausprobieren
#> 25                                                                                                                                                                                                                                                                                                                                                                                                           Scheint zu funktionieren!:P
#> 28                                                                                                                                                                                                                                                                                                                                                                                                            https://github.com/JuKo007
#> 29                                                                                                                                                                                                                                                                                                                                         Funktioniert! Gehen auch mehrere Links? https://github.com/JuKo007 https://github.com/JuKo007
#> 30                                                                                                                                                                                                                                                                                                                                                                                                                           Sehr schön!
#> 33                                                                                                                                                                                                                                                                                                                                                                                                           Live-Standort wird geteilt.
#> 37                                                                                                                                                                                                                                                                                                                                                                    Damit sollten sich WhatsApp Verläufe besser quantifizieren lassen!
#> 40                                                                                                                                                                                                                                                                                                                                                                                                           Klar, das klingt mega cool!
#> 48 Ich gebe hiermit meine informierte Einwilligung zur Spende von anonymisierte Metainformationen aus dieser Konversation für Forschungszwecke. Es werden keine personenbezogenen Daten über mich gespeichert. Mir ist bewusst, dass meine Daten für Forschungszwecke zum Thema Datenspendeverhalten in aggregierter Form für Publikationen genutzt werden. Meine Daten können auch anonymisert anderen Forschern bereitgestellt werden.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                Flat
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                               Hey
#> 11                                                                                                                                                                                                                                                                                                                                                                                                     Ist das hier ein Chatverlauf
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                             <NA>
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                               ja
#> 21                                                                                                                                                                                                                                                                                                                                                                                                            und Emojis gehen auch
#> 23                                                                                                                                                                                                                                                                                                                                                                                  Keine Ahnung lass uns das doch mal ausprobieren
#> 25                                                                                                                                                                                                                                                                                                                                                                                                       Scheint zu funktionieren P
#> 28                                                                                                                                                                                                                                                                                                                                                                                                                             <NA>
#> 29                                                                                                                                                                                                                                                                                                                                                                                            Funktioniert Gehen auch mehrere Links
#> 30                                                                                                                                                                                                                                                                                                                                                                                                                       Sehr schön
#> 33                                                                                                                                                                                                                                                                                                                                                                                                                             <NA>
#> 37                                                                                                                                                                                                                                                                                                                                                                Damit sollten sich WhatsApp Verläufe besser quantifizieren lassen
#> 40                                                                                                                                                                                                                                                                                                                                                                                                        Klar das klingt mega cool
#> 48 Ich gebe hiermit meine informierte Einwilligung zur Spende von anonymisierte Metainformationen aus dieser Konversation für Forschungszwecke Es werden keine personenbezogenen Daten über mich gespeichert Mir ist bewusst dass meine Daten für Forschungszwecke zum Thema Datenspendeverhalten in aggregierter Form für Publikationen genutzt werden Meine Daten können auch anonymisert anderen Forschern bereitgestellt werden
#>          TokVec          URL                    Media
#> 4           Hey           NA                     <NA>
#> 11 Ist, das....           NA                     <NA>
#> 15           NA           NA PTT-20180129-WA0025.opus
#> 19           ja           NA                     <NA>
#> 21 und, Emo....           NA                     <NA>
#> 23 Keine, A....           NA                     <NA>
#> 25 Scheint,....           NA                     <NA>
#> 28           NA https://....                     <NA>
#> 29 Funktion.... https://....                     <NA>
#> 30  Sehr, schön           NA                     <NA>
#> 33           NA           NA                     <NA>
#> 37 Damit, s....           NA                     <NA>
#> 40 Klar, da....           NA                     <NA>
#> 48 Ich, geb....           NA                     <NA>
#>                       Location Emoji EmojiDescriptions Smilies SystemMessage
#> 4                         <NA>    NA                NA      :)          <NA>
#> 11                        <NA>    NA                NA      NA          <NA>
#> 15                        <NA>    NA                NA      NA          <NA>
#> 19                        <NA>    😄      Grinning....      NA          <NA>
#> 21                        <NA>    NA                NA      NA          <NA>
#> 23                        <NA>    NA                NA      NA          <NA>
#> 25                        <NA>    NA                NA      NA          <NA>
#> 28                        <NA>    NA                NA      NA          <NA>
#> 29                        <NA>    NA                NA      NA          <NA>
#> 30                        <NA>    NA                NA      NA          <NA>
#> 33 Live-Standort wird geteilt.    NA                NA      NA          <NA>
#> 37                        <NA>    NA                NA      NA          <NA>
#> 40                        <NA>    NA                NA      NA          <NA>
#> 48                        <NA>    NA                NA      NA          <NA>
#>    TokCount TimeOrder DisplayOrder
#> 4         1         4            4
#> 11        5        11           11
#> 15        1        15           15
#> 19        1        19           19
#> 21        4        21           21
#> 23        8        23           23
#> 25        4        25           25
#> 28        1        28           28
#> 29        5        29           29
#> 30        2        30           30
#> 33        1        33           33
#> 37        8        37           37
#> 40        5        40           40
#> 48       51        50           48